+++ src/Utilities.Test/Utilities.Test.csproj
@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>net6.0;net5.0;netcoreapp3.1;net48;net462</TargetFrameworks>
+    <TargetFrameworks>netstandard2.0;net6.0;net5.0;netcoreapp3.1;net471;net462;net461</TargetFrameworks>
     <RootNamespace>Matr.Utilities.Test</RootNamespace>
     <EnableDefaultCompileItems>false</EnableDefaultCompileItems>
 
@@ -34,25 +34,22 @@
   <ItemGroup>
     <PackageReference Include="autofac" Version="6.2.0" />
     <PackageReference Include="MSTest.TestFramework" Version="2.2.5" />
-    <PackageReference Condition=" '$(TargetFramework)' == 'net48' " Include="system.text.json" Version="5.0.2" />
-    <PackageReference Condition=" '$(TargetFramework)' == 'net462' " Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageReference Condition=" '$(TargetFramework)' == 'netstandard2.0' OR '$(TargetFramework)' == 'net471' OR '$(TargetFramework)' == 'net462' OR '$(TargetFramework)' == 'net461' " Include="system.text.json" Version="5.0.2" />
   </ItemGroup>
 
   <ItemGroup>
     <Compile Include="TestBase.cs" />
-    <Compile Include="GenericFactory_base.cs" />
-    <Compile Include="Attributes\JsonTestMethodAttribute_base.cs" />
+    <Compile Include="Attributes\JsonTestMethodAttribute.cs" />
 
-    <Compile Condition=" '$(TargetFramework)' != 'net462' " Include="Attributes\JsonTestMethodAttribute.cs" />
-    <Compile Condition=" '$(TargetFramework)' == 'net462' " Include="Attributes\JsonTestMethodAttribute_462.cs" />
+    <Compile Include="GenericFactory_base.cs" />
   </ItemGroup>
 
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net6.0' OR '$(TargetFramework)' == 'net5.0' OR '$(TargetFramework)' == 'netcoreapp3.1'  ">
+  <ItemGroup Condition=" '$(TargetFramework)' == 'net6.0' OR '$(TargetFramework)' == 'net5.0' OR '$(TargetFramework)' == 'netcoreapp3.1' ">
     <Compile Include="GenericFactory.cs" />
   </ItemGroup>
 
 
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net462' OR '$(TargetFramework)' == 'net48' OR '$(TargetFramework)' == 'netstandard2.1' OR '$(TargetFramework)' == 'netstandard2.0'">
+  <ItemGroup Condition=" '$(TargetFramework)' == 'net471' OR '$(TargetFramework)' == 'net461' OR '$(TargetFramework)' == 'net462' OR '$(TargetFramework)' == 'net48' OR '$(TargetFramework)' == 'netstandard2.1' OR '$(TargetFramework)' == 'netstandard2.0' ">
     <Compile Include="GenericFactory_462.cs" />
   </ItemGroup>
 </Project>
